web/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-multipart==0.0.19
pandas==2.2.3
numpy==2.0.2
python-dotenv==1.0.1
curl_cffi>=0.5.0
loguru==0.7.3
